pull the overflow out of SendToPlan to a new TextButton
